tags/licenses fetch from <EditImage /> + other fixes from codereview
